Lots of crashes every day due to IabHelper crash
It seems that IabHelper is an old implementation and is not properly garbadge collected.
I don't have time to investigate, but crashing because of this is not needed. The typical described behavior seems to be that users view the products, go elsewhere and it crashes several seconds after when it's being garbage collected or something.
I plan on switching to j3k because the maintainers of this plugin are not active anymore, but this should at least reduce the crashes for now. 

REFs:
https://stackoverflow.com/questions/15456434/billing-api-v3-iabhelper-nullpointerexception
https://github.com/AlexDisler/cordova-plugin-inapppurchase/issues/129
https://github.com/AlexDisler/cordova-plugin-inapppurchase/issues/168